requirements_dev.txt
django-extensions==3.1.5
ipython==8.1.1

